refactor: retarget profile owner
