+++ tests/Integration/test_marketplace_router_user_shell.py
@@ -30,7 +30,7 @@ async def test_publish_agent_user_to_marketplace_uses_user_repo_not_member_repo(
 
     monkeypatch.setattr(marketplace_router.marketplace_client, "publish", lambda **kwargs: seen.update(kwargs) or {"ok": True})
     monkeypatch.setattr(
-        "backend.web.services.profile_service.get_profile",
+        "backend.profile.get_profile",
         lambda user=None: (
             (_ for _ in ()).throw(AssertionError("profile lookup must be user-scoped")) if user is None else {"name": user.display_name}
         ),
